Made the music lock make more sense
The music lock used to just be the "fade" lock.
This made things a bit weird since almost everything required it anyway, and at one point it just locked and unlocked it again to make sure the lock was open.
Making it the music lock means that every time the music is changed, it gets locked.
The code is much cleaner and simplier with it.